Refactor EncryptingFileSystem to Efsrpc
The DCERPC modules are mostly named after the well-known pipe that a
user would connect to.
