+++ lib/ruby_smb/dcerpc.rb
@@ -48,7 +48,7 @@ module RubySMB
     require 'ruby_smb/dcerpc/sec_trailer'
     require 'ruby_smb/dcerpc/dfsnm'
     require 'ruby_smb/dcerpc/icpr'
-    require 'ruby_smb/dcerpc/encrypting_file_system'
+    require 'ruby_smb/dcerpc/efsrpc'
     require 'ruby_smb/dcerpc/lsarpc'
     require 'ruby_smb/dcerpc/request'
     require 'ruby_smb/dcerpc/response'
