+++ lib/ruby_smb/dcerpc/efsrpc/efs_rpc_decrypt_file_srv_response.rb
@@ -1,6 +1,6 @@
 module RubySMB
   module Dcerpc
-    module EncryptingFileSystem
+    module Efsrpc
 
       # [3.1.4.2.6 Receiving an EfsRpcDecryptFileSrv Message (Opnum 5)](https://learn.microsoft.com/en-us/openspecs/windows_protocols/ms-efsr/043715de-caee-402a-a61b-921743337e78)
       class EfsRpcDecryptFileSrvResponse < BinData::Record
