+++ lib/ruby_smb/dcerpc/efsrpc/efs_rpc_encrypt_file_srv_request.rb
@@ -1,6 +1,6 @@
 module RubySMB
   module Dcerpc
-    module EncryptingFileSystem
+    module Efsrpc
 
       # [3.1.4.2.5 EfsRpcEncryptFileSrv (Opnum 4)](https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-efsr/0d599976-758c-4dbd-ac8c-c9db2a922d76)
       class EfsRpcEncryptFileSrvRequest < BinData::Record
