+++ lib/ruby_smb/smb1/pipe.rb
@@ -35,7 +35,7 @@ module RubySMB
         when 'cert', '\\cert'
           extend RubySMB::Dcerpc::Icpr
         when 'efsrpc', '\\efsrpc'
-          extend RubySMB::Dcerpc::EncryptingFileSystem
+          extend RubySMB::Dcerpc::Efsrpc
         end
         super(tree: tree, response: response, name: name)
       end
